Changes (to Customer):
-distinguished first/last name as specified by specs
-added a toString() method (THIS IS NOT A REPORT!)
-removed the list of transactions and the debt() method

I think it would be best to have customer only store data, and handle the transactions/receivables through a coordinating class that contains both the customers and the transactions. This is for two reasons: The first being that we need to have a class like that anyways, so we might as well put the calculations there. The second reason is that we want to store as little as possible in Customer so we can avoid having to keep updated separate lists of transactions for every customer and instead handle only one main list (KISS). Also, it fits the specs better.

Another thing to note is that, given the current implementation, Transactions have their own report() function while Customer's report will be handled by the coordinating class. My justification for this is that a Transaction's report only requires data stored within the class, while a Customer's report needs to also access data from other classes (for displaying order history, for example). It also gives us more modularity within the Transactions. However, I don't really like this assymmetry, and I think it would be more consistent to have the coordinating class handle all reports.