Added diagram of architecture
Shop contains the data structures (list of people, prices, etc), as well as functions that directly correspond to the commands.
IOHandler (which should probably be called InputHandler since it doesn't handle output at all) contains an instance of Shop. IOHandler reads input, translates it to the proper Shop function, and calls it. Most output is delivered directly by shop.