added printworp()
printworp() prints out the revenue for each week of business from orders. Each week also shows all of its component days and the amount of order revenue they got. This is only to see if my program actually works in terms of calculating revenues; in the future it would be better to have a graph for each week. Just now I realized that I severely messed up in adding the option to see revenues of component days, the program should only look at the revenue from a whole week and compare it with that of other weeks.