helpers for printworp()
I added too much info. For the Week class the 2nd-6th day properties can be removed. 1st and 7th are still necessary because it shows the range of the week. the whole thing in the weeklyOrderRevenueReport with the order revenue of each day is unnecessary, so cut those out. consequently the helper ordersOfDate can also be cut.